feat: new, snappier, modal setup (#185)
* feat: add fallback to metadata requests

* refactor: new, snappier, modal setup
