+++ src/app/(pages)/(home)/@modal/(.)nominate/[walletId]/loading.tsx
@@ -1,0 +1,3 @@
+import NominateBuilderLoading from "@/app/_modals/nominate-builder/loading";
+
+export default NominateBuilderLoading;
