+++ src/app/(pages)/(home)/@modal/(.)nominate/[walletId]/not-found.tsx
@@ -1,0 +1,3 @@
+import NominatedBuilderNotFound from "@/app/_modals/nominate-builder/not-found";
+
+export default NominatedBuilderNotFound;
