--- src/app/(pages)/(home)/default.tsx
@@ -1,5 +1,0 @@
-import { HomePageComponent } from "./component";
-
-export default function HomePageLoading() {
-  return <HomePageComponent loading />;
-}
