+++ src/app/(pages)/(home)/layout.tsx
@@ -1,0 +1,14 @@
+export default async function HomePageLayout({
+  children,
+  modal,
+}: {
+  children: React.ReactNode;
+  modal: React.ReactNode;
+}) {
+  return (
+    <>
+      {children}
+      {modal}
+    </>
+  );
+}
