--- src/app/@modal/(home)/nominate/[walletId]/loading.tsx
@@ -1,3 +1,0 @@
-import { default as NominateBuilderLoading } from "@/app/@modal/_modals/nominate-builder/loading";
-
-export default NominateBuilderLoading;
