--- src/app/@modal/(home)/nominate/[walletId]/not-found.tsx
@@ -1,3 +1,0 @@
-import { default as NominatedBuilderNotFound } from "@/app/@modal/_modals/nominate-builder/not-found";
-
-export default NominatedBuilderNotFound;
