--- src/app/@modal/(home)/nominate/[walletId]/page.tsx
@@ -1,4 +1,0 @@
-import { default as NominateBuilder } from "@/app/@modal/_modals/nominate-builder/page";
-
-export const maxDuration = 60;
-export default NominateBuilder;
