--- src/app/@modal/page.tsx
@@ -1,3 +1,0 @@
-export default function Nothing() {
-  return null;
-}
