+++ src/app/api/search/route.ts
@@ -1,14 +1,12 @@
 import { z } from "zod";
 import { searchBuilders } from "@/app/_api/functions/search-builders";
 import { restApiHandler } from "@/app/_api/helpers/rest-api-handler";
-import { rollbarError, rollbarInfo } from "@/services/rollbar";
 
 const searchParamsSchema = z.object({
   query: z.string().min(1),
   domain: z.string(),
 });
 
-export const maxDuration = 60;
 export const GET = restApiHandler(async (request) => {
   const searchParams = searchParamsSchema.parse({
     query: request.nextUrl.searchParams.get("query"),
@@ -19,4 +17,5 @@ export const GET = restApiHandler(async (request) => {
   return result;
 });
 
+export const maxDuration = 60;
 export const dynamic = "force-dynamic";
