+++ src/app/layout.tsx
@@ -1,7 +1,8 @@
+import { Metadata } from "next";
 import { headers } from "next/headers";
 import "@fontsource-variable/bricolage-grotesque";
 import "@fontsource-variable/inter";
-import { Box } from "@mui/joy";
+import { Box, Stack } from "@mui/joy";
 import { Toaster } from "sonner";
 import { cookieToInitialState } from "wagmi";
 import { Web3ModalProvider } from "@/app/_providers/web-3-modal-provider";
@@ -11,12 +12,36 @@ import { Header } from "@/shared/components/header";
 import { ThemeRegistry } from "@/shared/theme/theme-registry";
 import { AuthenticationProvider } from "./_providers/authentication-provider";
 
+const description = [
+  "BUILD is a meme and a social game designed to reward builders via",
+  "onchain nominations.",
+].join(" ");
+
+export const metadata: Metadata = {
+  title: "BUILD",
+  description: description,
+  openGraph: {
+    title: "BUILD",
+    description: description,
+    type: "website",
+    url: "https://build.top",
+    images: [
+      "https://build-top-images.s3.eu-west-2.amazonaws.com/BUILD-thumbnail.jpg?1",
+    ],
+  },
+  twitter: {
+    title: "BUILD",
+    description: description,
+    images: [
+      "https://build-top-images.s3.eu-west-2.amazonaws.com/BUILD-thumbnail.jpg?1",
+    ],
+  },
+};
+
 export default function RootLayout({
   children,
-  modal,
 }: Readonly<{
   children: React.ReactNode;
-  modal: React.ReactNode;
 }>) {
   const initialState = cookieToInitialState(config, headers().get("cookie"));
 
@@ -79,10 +104,9 @@ export default function RootLayout({
           <Web3ModalProvider initialState={initialState}>
             <AuthenticationProvider>
               <Header />
-              {children}
+              <Stack sx={{ minHeight: "80vh" }}>{children}</Stack>
               <Footer />
               <Toaster richColors closeButton />
-              {modal}
             </AuthenticationProvider>
           </Web3ModalProvider>
         </Box>
@@ -90,5 +114,3 @@ export default function RootLayout({
     </html>
   );
 }
-
-export const dynamic = "force-dynamic";
